// frontend-legacy/src/app/assets-discover/shared/types/asset-filter.type.ts
export class AssetFilterType {
  alive?: boolean;
  discoveredEndDate?: string;
  discoveredInitDate?: string;
  assetIpMacName?: string;
  size?: number;
  sort?: string;
  probe?: string[];
  page?: number;
  openPorts?: string[];
  os?: string[];
  severity?: string[];
  status?: string[];
  type?: string[];
  groups?: string[];
  alias?: string[];
  registeredMode?: string;
  dataTypes?: string[];
}